requirements.txt
alembic==1.0.7
astroid==2.2.2
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
Click==7.0
dialogflow==0.5.2
Flask==1.0.2
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
google-api-core==1.8.0
google-auth==1.6.3
googleapis-common-protos==1.5.8
grpcio==1.19.0
idna==2.8
isort==4.3.10
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
Mako==1.0.7
MarkupSafe==1.1.1
mccabe==0.6.1
protobuf==3.7.0
psycopg2-binary==2.7.7
pyasn1==0.4.5
pyasn1-modules==0.2.4
pylint==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.4
pytz==2018.9
requests==2.21.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.2.18
typed-ast==1.3.1
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
